refactor: 컨테이너 및 설정 파일 위치 변경
